requirements.txt
fastapi==0.61.2
imageai==2.1.5
Keras==2.3.1
kiwisolver==1.3.1
matplotlib==3.3.3
numpy
opencv-python==4.1.2.30
Pillow==8.0.1
pytesseract==0.3.6
python-multipart==0.0.5
scikit-image==0.17.2
scipy 
starlette==0.13.6
tensorboard==2.0.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
urllib3==1.26.2
uvicorn==0.12.2
gunicorn==20.0.4
shutil